Move `character.profile.Guild` to `character.Guild`
