+++ src/main/kotlin/character/search/ScrapeCharacterSearch.kt
@@ -3,7 +3,7 @@ package cloud.drakon.ktlodestone.character.search
 import cloud.drakon.ktlodestone.character.grandcompany.GrandCompany
 import cloud.drakon.ktlodestone.character.grandcompany.GrandCompanyName
 import cloud.drakon.ktlodestone.character.grandcompany.GrandCompanyRank
-import cloud.drakon.ktlodestone.character.profile.Guild
+import cloud.drakon.ktlodestone.character.Guild
 import cloud.drakon.ktlodestone.selectors.character.search.CharacterSearchSelectors
 import cloud.drakon.ktlodestone.world.DataCenter
 import cloud.drakon.ktlodestone.world.World
